Fix demo issues
